Move rsspp::Exception to a separate file
